@@ -7,6 +7,7 @@
     "url": "https://github.com/soketi/soketi.git"
   },
   "main": "dist/index.js",
+  "type": "module",
   "keywords": [
     "laravel",
     "socket.io",
@@ -22,6 +23,9 @@
     "esversion": 9
   },
   "dependencies": {
+    "@chainsafe/libp2p-noise": "^8.0.1",
+    "@libp2p/mplex": "^5.2.4",
+    "@libp2p/tcp": "^3.1.2",
     "arraybuffer-to-string": "^1.0.2",
     "async": "^3.2.4",
     "aws-sdk": "^2.1194.0",
@@ -35,7 +39,6 @@
     "knex": "^1.0.7",
     "mysql": "^2.18.1",
     "mysql2": "^2.3.3",
-    "node-discover": "^1.2.1",
     "pg": "^8.7.1",
     "pm2": "^5.2.0",
     "prom-client": "^14.0.1",
@@ -72,7 +75,7 @@
     "tcp-port-used": "^1.0.2",
     "ts-jest": "^26.5.6",
     "tslib": "^2.3.1",
-    "typescript": "^4.5.4"
+    "typescript": "^4.8.4"
   },
   "scripts": {
     "build": "node node_modules/typescript/bin/tsc",
