@@ -1,6 +1,6 @@
 #! /usr/bin/env node
 
-const { Cli } = require('./../dist/cli/cli');
+import { Cli } from './../dist/cli/cli.js';
 
 process.title = 'soketi-server';
 
