@@ -6,8 +6,8 @@ import ws from 'k6/ws';
  *
  * 1. Run the servers:
  *
- * SOKETI_PORT=6001 SOKETI_ADAPTER_DRIVER=cluster SOKETI_RATE_LIMITER_DRIVER=cluster bin/server.js start
- * SOKETI_PORT=6002 SOKETI_ADAPTER_DRIVER=cluster SOKETI_RATE_LIMITER_DRIVER=cluster bin/server.js start
+ * SOKETI_PORT=6001 SOKETI_ADAPTER_DRIVER=local SOKETI_RATE_LIMITER_DRIVER=local bin/server.js start
+ * SOKETI_PORT=6002 SOKETI_ADAPTER_DRIVER=local SOKETI_RATE_LIMITER_DRIVER=local bin/server.js start
  *
  * 2. Run the PHP senders based on the amount of messages per second you want to receive.
  *    The sending rate influences the final benchmark.
@@ -37,7 +37,7 @@ if (['mysql', 'postgres', 'dynamodb'].includes(__ENV.APP_MANAGER_DRIVER)) {
 }
 
 // Horizontal drivers take additional time to communicate with other nodes.
-if (['redis', 'cluster'].includes(__ENV.ADAPTER_DRIVER)) {
+if (['redis'].includes(__ENV.ADAPTER_DRIVER)) {
     maxP95 += 100;
     maxAvg += 100;
 }
