@@ -1,5 +1,5 @@
 #! /usr/bin/env node
 
-require('./../dist/cli');
+import './../dist/cli/cli.js';
 
 process.title = 'soketi-server';
