@@ -7,7 +7,6 @@ import { HttpHandler } from './http-handler';
 import { HttpRequest, HttpResponse, TemplatedApp } from 'uWebSockets.js';
 import { Log } from './log';
 import { Metrics, MetricsInterface } from './metrics';
-import { Node } from './node';
 import { Options } from './options';
 import { Queue } from './queues/queue';
 import { QueueInterface } from './queues/queue-interface';
@@ -19,7 +18,6 @@ import { WebhookSender } from './webhook-sender';
 import { WebSocket } from 'uWebSockets.js';
 import { WsHandler } from './ws-handler';
 
-const Discover = require('node-discover');
 const queryString = require('query-string');
 const uWS = require('uWebSockets.js');
 
@@ -41,9 +39,6 @@ export class Server {
                 },
                 clusterMode: false,
                 shardMode: false,
-            },
-            cluster: {
-                requestsTimeout: 5_000,
             },
         },
         appManager: {
@@ -95,19 +90,6 @@ export class Server {
         channelLimits: {
             maxNameLength: 200,
             cacheTtl: 3600,
-        },
-        cluster: {
-            hostname: '0.0.0.0',
-            helloInterval: 500,
-            checkInterval: 500,
-            nodeTimeout: 2000,
-            masterTimeout: 2000,
-            port: 11002,
-            prefix: '',
-            ignoreProcess: true,
-            broadcast: '255.255.255.255',
-            unicast: null,
-            multicast: null,
         },
         cors: {
             credentials: true,
@@ -302,16 +284,6 @@ export class Server {
     public pm2 = false;
 
     /**
-     * The list of nodes in the current private network.
-     */
-    public nodes: Map<string, Node> = new Map<string, Node>();
-
-    /**
-     * The Discover instance.
-     */
-    public discover: typeof Discover;
-
-    /**
      * Initialize the server.
      */
     constructor(options = {}) {
@@ -331,64 +303,61 @@ export class Server {
     async start(callback?: CallableFunction) {
         Log.br();
 
-        this.configureDiscovery().then(() => {
-            this.initializeDrivers().then(() => {
+        this.initializeDrivers().then(() => {
+            if (this.options.debug) {
+                console.dir(this.options, { depth: 100 });
+            }
 
+            this.wsHandler = new WsHandler(this);
+            this.httpHandler = new HttpHandler(this);
+
+            if (this.options.debug) {
+                Log.info('📡 soketi initialization....');
+                Log.info('⚡ Initializing the HTTP API & Websockets Server...');
+            }
+
+            let server: TemplatedApp = this.shouldConfigureSsl()
+                ? uWS.SSLApp({
+                    key_file_name: this.options.ssl.keyPath,
+                    cert_file_name: this.options.ssl.certPath,
+                    passphrase: this.options.ssl.passphrase,
+                    ca_file_name: this.options.ssl.caPath,
+                })
+                : uWS.App();
+
+            let metricsServer: TemplatedApp = uWS.App();
+
+            if (this.options.debug) {
+                Log.info('⚡ Initializing the Websocket listeners and channels...');
+            }
+
+            this.configureWebsockets(server).then(server => {
                 if (this.options.debug) {
-                    console.dir(this.options, { depth: 100 });
+                    Log.info('⚡ Initializing the HTTP webserver...');
                 }
 
-                this.wsHandler = new WsHandler(this);
-                this.httpHandler = new HttpHandler(this);
+                this.configureHttp(server).then(server => {
+                    this.configureMetricsServer(metricsServer).then(metricsServer => {
+                        metricsServer.listen(this.options.metrics.host, this.options.metrics.port, metricsServerProcess => {
+                            this.metricsServerProcess = metricsServerProcess;
 
-                if (this.options.debug) {
-                    Log.info('📡 soketi initialization....');
-                    Log.info('⚡ Initializing the HTTP API & Websockets Server...');
-                }
+                            server.listen(this.options.host, this.options.port, serverProcess => {
+                                this.serverProcess = serverProcess;
 
-                let server: TemplatedApp = this.shouldConfigureSsl()
-                    ? uWS.SSLApp({
-                        key_file_name: this.options.ssl.keyPath,
-                        cert_file_name: this.options.ssl.certPath,
-                        passphrase: this.options.ssl.passphrase,
-                        ca_file_name: this.options.ssl.caPath,
-                    })
-                    : uWS.App();
+                                Log.successTitle('🎉 Server is up and running!');
+                                Log.successTitle(`📡 The Websockets server is available at 127.0.0.1:${this.options.port}`);
+                                Log.successTitle(`🔗 The HTTP API server is available at http://127.0.0.1:${this.options.port}`);
+                                Log.successTitle(`🎊 The /usage endpoint is available on port ${this.options.metrics.port}.`);
 
-                let metricsServer: TemplatedApp = uWS.App();
+                                if (this.options.metrics.enabled) {
+                                    Log.successTitle(`🌠 Prometheus /metrics endpoint is available on port ${this.options.metrics.port}.`);
+                                }
 
-                if (this.options.debug) {
-                    Log.info('⚡ Initializing the Websocket listeners and channels...');
-                }
+                                Log.br();
 
-                this.configureWebsockets(server).then(server => {
-                    if (this.options.debug) {
-                        Log.info('⚡ Initializing the HTTP webserver...');
-                    }
-
-                    this.configureHttp(server).then(server => {
-                        this.configureMetricsServer(metricsServer).then(metricsServer => {
-                            metricsServer.listen(this.options.metrics.host, this.options.metrics.port, metricsServerProcess => {
-                                this.metricsServerProcess = metricsServerProcess;
-
-                                server.listen(this.options.host, this.options.port, serverProcess => {
-                                    this.serverProcess = serverProcess;
-
-                                    Log.successTitle('🎉 Server is up and running!');
-                                    Log.successTitle(`📡 The Websockets server is available at 127.0.0.1:${this.options.port}`);
-                                    Log.successTitle(`🔗 The HTTP API server is available at http://127.0.0.1:${this.options.port}`);
-                                    Log.successTitle(`🎊 The /usage endpoint is available on port ${this.options.metrics.port}.`);
-
-                                    if (this.options.metrics.enabled) {
-                                        Log.successTitle(`🌠 Prometheus /metrics endpoint is available on port ${this.options.metrics.port}.`);
-                                    }
-
-                                    Log.br();
-
-                                    if (callback) {
-                                        callback(this);
-                                    }
-                                });
+                                if (callback) {
+                                    callback(this);
+                                }
                             });
                         });
                     });
@@ -542,78 +511,6 @@ export class Server {
      */
     protected url(path: string): string {
         return this.options.pathPrefix + path;
-    }
-
-    /**
-     * Get the cluster prefix name for discover.
-     */
-    clusterPrefix(channel: string): string {
-        if (this.options.cluster.prefix) {
-            channel = this.options.cluster.prefix + '#' + channel;
-        }
-
-        return channel;
-    }
-
-    /**
-     * Configure the private network discovery for this node.
-     */
-    protected configureDiscovery(): Promise<void> {
-        return new Promise(resolve => {
-            this.discover = Discover(this.options.cluster, () => {
-                this.nodes.set('self', this.discover.me);
-
-                this.discover.on('promotion', () => {
-                    this.nodes.set('self', this.discover.me);
-
-                    if (this.options.debug) {
-                        Log.discoverTitle('Promoted from node to master.');
-                        Log.discover(this.discover.me);
-                    }
-                });
-
-                this.discover.on('demotion', () => {
-                    this.nodes.set('self', this.discover.me);
-
-                    if (this.options.debug) {
-                        Log.discoverTitle('Demoted from master to node.');
-                        Log.discover(this.discover.me);
-                    }
-                });
-
-                this.discover.on('added', (node: Node) => {
-                    this.nodes.set('self', this.discover.me);
-                    this.nodes.set(node.id, node);
-
-                    if (this.options.debug) {
-                        Log.discoverTitle('New node added.');
-                        Log.discover(node);
-                    }
-                });
-
-                this.discover.on('removed', (node: Node) => {
-                    this.nodes.set('self', this.discover.me);
-                    this.nodes.delete(node.id);
-
-                    if (this.options.debug) {
-                        Log.discoverTitle('Node removed.');
-                        Log.discover(node);
-                    }
-                });
-
-                this.discover.on('master', (node: Node) => {
-                    this.nodes.set('self', this.discover.me);
-                    this.nodes.set(node.id, node);
-
-                    if (this.options.debug) {
-                        Log.discoverTitle('New master.');
-                        Log.discover(node);
-                    }
-                });
-
-                resolve();
-            });
-        });
     }
 
     /**
