@@ -1,6 +1,5 @@
 import async from 'async';
 import { Log } from '../src/log';
-import { PusherApiMessage } from '../src/message';
 import { Server } from './../src/server';
 import { v4 as uuidv4 } from 'uuid';
 
@@ -33,21 +32,19 @@ export class Utils {
             tcpPortUsed.waitUntilFree(3001, 500, 5 * 1000),
             tcpPortUsed.waitUntilFree(9601, 500, 5 * 1000),
             tcpPortUsed.waitUntilFree(11002, 500, 5 * 1000),
+            tcpPortUsed.waitUntilFree(115353, 500, 5 * 1000),
         ]);
     }
 
     static newServer(options = {}, callback): any {
         options = {
-            'cluster.prefix': uuidv4(),
             'adapter.redis.prefix': uuidv4(),
             'appManager.array.apps.0.maxBackendEventsPerSecond': 200,
             'appManager.array.apps.0.maxClientEventsPerSecond': 200,
             'appManager.array.apps.0.maxReadRequestsPerSecond': 200,
             'metrics.enabled': true,
             'appManager.mysql.useMysql2': true,
-            'cluster.port': parseInt((Math.random() * (20000 - 10000) + 10000).toString()), // random: 10000-20000
             'appManager.dynamodb.endpoint': 'http://127.0.0.1:8000',
-            'cluster.ignoreProcess': false,
             'webhooks.batching.enabled': false,
             'webhooks.batching.duration': 50,
             'database.redis.enableOfflineQueue': true,
@@ -87,8 +84,6 @@ export class Utils {
         return this.newServer({
             // Make sure the same prefixes exists so that they can communicate
             'adapter.redis.prefix': server.options.adapter.redis.prefix,
-            'cluster.prefix': server.options.cluster.prefix,
-            'cluster.port': server.options.cluster.port,
             ...options,
         }, callback);
     }
